Code cleanup, Refactoring
