+++ frontend/src/app/app.component.html
@@ -1,0 +1,9 @@
+<mat-toolbar>
+  <button mat-button routerLink="/">Home</button>
+  <span style="flex: 1 1 auto"></span>
+  <button mat-button routerLink="/users"> Users </button>
+  <button mat-button *ngIf="!authService.isAuthenticated" routerLink="/register"> Register </button>
+  <button mat-button *ngIf="!authService.isAuthenticated" routerLink="/login"> Login </button>
+  <button mat-button *ngIf="authService.isAuthenticated" (click)="authService.logout()"> Logout </button>
+</mat-toolbar>
+<router-outlet></router-outlet>
