+++ frontend/src/app/app.component.ts
@@ -1,22 +1,13 @@
 import { Component } from '@angular/core';
-import { AuthService } from './auth.service';
+import { AuthService } from './auth/auth.service';
 
 @Component({
-  selector: 'app-root',
-  template: `
-  <mat-toolbar>
-  <button mat-button routerLink="/">Home</button>
-  <span style="flex: 1 1 auto"></span>
-  <button mat-button  routerLink="/users"> Users </button>
-  <button mat-button *ngIf="!authService.isAuthenticated" routerLink="/register"> Register </button>
-  <button mat-button *ngIf="!authService.isAuthenticated" routerLink="/login"> Login </button>
-  <button mat-button *ngIf="authService.isAuthenticated" (click)="authService.logout()"> Logout </button>
-  </mat-toolbar>
-  <router-outlet></router-outlet>`,
-  styleUrls: ['./app.component.css'],
+	selector: 'app-root',
+	templateUrl: 'app.component.html',
+	styleUrls: ['./app.component.css'],
 })
 export class AppComponent {
-  constructor(private authService: AuthService) {}
+	constructor(private authService: AuthService) {}
 
-  title = 'app';
+	title = 'app';
 }
