+++ frontend/src/app/posts/messages.component.html
@@ -1,0 +1,4 @@
+<div *ngFor="let message of apiService.messages">
+  <mat-card>{{message.msg}}</mat-card>
+
+</div>
