+++ frontend/src/app/user/login.component.html
@@ -1,0 +1,18 @@
+<mat-card>
+  <mat-card-header>
+    <mat-card-title>
+      <h4>Login New User</h4>
+    </mat-card-title>
+  </mat-card-header>
+  <mat-card-content>
+    <form>
+      <mat-form-field class="example-full-width">
+        <input [(ngModel)]="loginData.email" name="email" matInput placeholder="email" type="email">
+      </mat-form-field>
+      <mat-form-field class="example-full-width">
+        <input [(ngModel)]="loginData.pswd" name="password" matInput placeholder="password" type="password">
+      </mat-form-field>
+      <button (click)="post()" mat-raised-button color="primary">Login</button>
+    </form>
+  </mat-card-content>
+</mat-card>
