+++ frontend/src/app/user/profile.component.html
@@ -1,0 +1,24 @@
+<mat-card>
+  <mat-card-header>
+    <mat-card-title>
+      <h4>Profile</h4>
+    </mat-card-title>
+  </mat-card-header>
+  <mat-card-content>
+    <mat-list>
+      <mat-list-item>Email: {{profile?.email}}</mat-list-item>
+      <mat-list-item>Name: {{profile?.name}}</mat-list-item>
+      <mat-list-item>Description: {{profile?.description}}</mat-list-item>
+    </mat-list>
+  </mat-card-content>
+</mat-card>
+<mat-card>
+  <mat-card-header>
+    <mat-card-title>
+      <h4>Posts</h4>
+    </mat-card-title>
+  </mat-card-header>
+  <mat-card-content>
+    <messages></messages>
+  </mat-card-content>
+</mat-card>
