+++ frontend/src/app/user/users.component.html
@@ -1,0 +1,4 @@
+<div *ngFor="let user of apiService.users">
+  <mat-card [routerLink]="['/profile', user._id]" style="cursor: pointer;">{{user.name}}</mat-card>
+
+</div>
